Optionally use system installation of GLFW
When using Wayland, there's a reason for this change.
GLFW often throws an error like "Can't focus the window",
so I use minecraft-wayland-glfw from Fedora COPR to fix this.
And GLFW upstream of course haven't got this patch.

Sadly, didn't fix the bug. But will keep trying to run it :^)