Add TRUERPG_RES_DIR_PREFIX define for use when loading resources
By default, cmake sets this variable to ".." (to not break
compatibility with the current development environment (which sucks
imho)). But in the code it will be defined as "../res".

I replaced every occurency of ".." in the code with 'TRUERPG_RES_DIR
"<here goes the path of the resource in the res folder>"'. That basically
means that we can install the TrueRPG into any directory we'll want and
it'll not depend on ".." folder constraint.

This patch is made to make TrueRPG great again and to distribute it in
package managers, such as Portage. :^)